// models/view-models/coinone-price-model.ts
export interface ICoinonePriceModel {
    result: string;
    volume: number;
    last: number;
    yesterday_last: number;
    timestamp: number;
    yesterday_low: number;
    high: number;
    currency: string;
    low: number;
    yesterday_first: number;
    yesterday_volume: number;
    yesterday_high: number;
    first: number;
}